[primer] Add a custom prelude for `primer`.
This custom prelude, named Foreword, is based on Protolude. In this
initial version, we re-export most of Protolude, but we elide a few
types and functions whose names clash with types and functions we use
in Primer. We're unlikely to use these particular elided types &
functions very often, so this seems like the right trade-off. (In a
few cases, I've renamed our own functions or local variables to avoid
clashes, when the Protolude versions seemed generally more useful
and/or conventional.)

When we do elide a function, we also elide related functions. For
example, Primer exports a function named `check`, and we use it
extensively. This name conflicts with `Control.Monad.STM.check`, which
is exported unqualified from Protolude, and therefore we elide `check`
from Protolude's exports. But as a user of Foreword, it might be
confusing if you could use, say, `atomically`, also from STM, but not
`check`. Therefore, we do not re-export *any* STM functions from
Protolude. You would therefore need to import
`Control.Monad.STM` (possibly qualified) if you wanted to use any STM
functions.

Note that there is still a bit of work to do before we're taking full
advantage of Protolude. Specifically, we make use of `error` and
`head` in a few places in library code, and we should not be doing
that. (We never want Primer to crash or throw a generic exception,
even if those cases should "never happen.") In this commit, I've
renamed the uses of `head` to `unsafeHead`. Protolude does re-export
`error`, but it does so with a warning, which, while normally helpful,
in our case would cause CI to fail, so we don't take advantage of
this, and `error` is now imported from `GHC.Err`, for the moment.

Other notes:

* There are certainly other things we might want to add to our custom
prelude, but one step at a time. (`Optics` seems like an obvious
candidate to add next.)

* I believe that, for the most part, no types have changed in this
commit, and I've not significantly rewritten any code, so these
changes should be more or less semantically neutral.

* Protolude famously does not export `String`, and in a few places
I've had to import it from `Data.String`. We should probably replace
those uses with `Text` for better performance & consistency with the
rest of the code base.

* `writeHieFiles` in our Nix config breaks on `mtl-compat-0.2.2`,
which is pulled in by `protolude`. For now, I've disabled that option
just for this package. `mtl-compat` probably doesn't generate HIE
files, anyway. See:

https://github.com/input-output-hk/haskell.nix/issues/1242