Add weeder to the shell
Notice that the configuration sets just `Main.main` functions as roots,
and not any functions exported from libraries.  This works pretty well
when running over the whole project, since anything exported from a
library should be used in some testsuite or executable in this
repository.

It is not yet hooked up to CI.

We need to ask cabal to add hie files so weeder works in the shell.
Unfortunately haskell.nix does not pick this option up, so we need to
keep the writeHieFiles option in flake.nix also.