[refactor] Move some definitions out to new `Primer.Name.Fresh` module
This untangles our module dependency tree a little, in particular removing the dependency of `Question` on `Action`. In the next commit, we will introduce the inverse dependency.

Requires generalising the type signatures of `mkFreshName` and `mkFreshNameTy`. All this does is remove some redundant constraints coming from the `ActionM` synonym. Other than this, the contents of `Fresh.hs` (except the module header) come verbatim from `Action.hs`