[service] Paginate the getSessionsList output
We do not return anything other than the appropriate page of sessions.
In particular there is no metadata describing whether this is the whole
list, or the last page or how many total items there are.

We now use mkPositive from code which is not a type class instance, so
can revert the change to weeder.dhall