[service] Improve error message for Positive parse failures
Previously we relied on the 'Natural' instance and added an extra check
that the number was positive. This gave errors
- "a": Error parsing query parameter pageSize failed: could not parse: `a' (input does not start with a digit)
- "-1": Error parsing query parameter pageSize failed: underflow: -1 (should be a non-negative integer)
- "0": Error parsing query parameter pageSize failed: Non-positive value: 0

Note that the error for '-1' suggests the input could be '0', but that
will not be accepted.

We now rely on the 'Integer' instance and add a check that the number
was positive. The errors are now
- "a": Error parsing query parameter pageSize failed: could not parse: `a' (input does not start with a digit)
- "-1": Error parsing query parameter pageSize failed: Non-positive value: -1
- "0": Error parsing query parameter pageSize failed: Non-positive value: 0