Rework pagination setup
This is preparation for actually doing the pagination at the DB layer.
We do not know the total number of items before doing a DB operation,
which does not fit well with the pagination library. Whilst we could
write 'listSessions' to return the total and a monadic action to extract
the page, we prefer to be more atomic and directly return the total
number of items and the current page. This means we need to move away
from the pagination library.

Since we are moving away from the library, we also change 'Positive' to
be backed by an 'Int', which fits better with the rest of the codebase.