[selda] Actually use the DB to extract a page
Previously we extracted every session to a list and then extracted the
relevent page in a second stage.

Note that, in general, offset+limit pagination can be inefficient with
very large tables. We do not expect our tables to be large enough for
this to matter, so we prefer the simplicity currently.

We should also be aware of the surprising performance of 'count' in
PostgreSQL:
  https://www.postgresql.org/docs/14/functions-aggregate.html
says
                              Note
    Users accustomed to working with other SQL database management
    systems might be disappointed by the performance of the count
    aggregate when it is applied to the entire table. A query like:
      SELECT count(*) FROM sometable;
    will require effort proportional to the size of the table:
    PostgreSQL will need to scan either the entire table or the entirety
    of an index that includes all rows in the table.
We similarly expect that this will not be an issue for our expected
database sizes.