chore: Bump dependencies.
This commit bumps both Nix and Haskell dependencies.

Note that the latest Weeder must be built with GHC 9. As Weeder uses
GHC to do its analysis, and we're not yet on GHC 9, we pin Weeder to
the last version that can be built with GHC 8.10, to minimize any risk
of incompatibility.