Change API to serialize IDs and Names plainly
i.e. we now have '1' and '"foo"' instead of
'{unID:1}' and '{unName:"foo"}'

Almost all of this commit is testcase churn because of this change.`

The change to ToJSONKey enables us to output 'Map ID _' from our API
(whilst having correct OpenAPI3 documentation), since the map will be
serialized as an object, where the keys are 'show' of the IDs. However,
we decide to keep our list-based representation, as documented in a
comment. Note that this only applies to the "new api", and not to the
serialization test fixtures.