Add support for primitive types and functions
For now, the only such type is `Char`, and the set of functions is small and fairly arbitrary (taken from https://github.com/hackworthltd/primer/issues/199).

There is support for reducing applications of primitive functions using either evaluator, so long as they are fully applied and all arguments fully evaluated first. Tests of this behaviour are included.