refactor: Don't expose MonadState internals of MonadFresh instances
There is no use of these instances in the codebase. I know of no reason
one would want to monkey about in these internals, so let's remove the
foot-gun. Further, since MonadState has a fundep and lifts through many
common monad transformers, this useless instances removes the
possibility of having a more useful MonadState instance on a wrapped
type in the future.