Split `Def` in to `PrimDef` and `ASTDef`
Changes in this commit, outside of `Core.hs`, are trivial, mostly just consisting of renaming the old `Def` to `ASTDef`. We will make use of the ability to treat both types of definitions more uniformly in the next commit. This split is intended to make the changes easier to review, which would be difficult if the two commits were squashed.

In passing, we happen to remove all uses of the partial function `Map.!` in the project (except in tests), seeing as we had to modify that code anyway (see `App.lookupASTDef`).