feat: Scripts to create & run a Docker container with PostgreSQL.
This uses `colima` to create the VMs used by the `docker` executable,
so the user need not have Docker installed, though it should co-exist
with Docker (via the use of Docker contexts) if they do.

The configuration makes use of a persistent Docker volume, so the
database should survive a container restart (and even a PostgreSQL
version upgrade).

This is all pretty imperative, but I don't know of a good way around
that which is also portable.

Docs are forthcoming in a subsequent commit.