feat: Add a Rel8 PostgreSQL adapter.
This will replace the Selda adapter in a subsequent commit. Tests are
forthcoming.

Notes:

- Rel8 communicates with the PostgreSQL database via Hasql, and Hasql
bubbles up database exceptions to the caller in-band via `Either`. We
want to handle these out-of-band, so we throw them to the creator of
the database thread, who presumably knows more about Hasql than core
Primer.

- Unlike Selda, Rel8 does not include support for creating tables.
This is possible in Hasql, but most best SQL practice seems to
recommend that schemas are created, migrated, etc. with a separate
tool. Therefore, I have not added an `initialize` action here to
mirror the one in our Selda adapter. Instead, we'll add support for
schema management in separate work.