feat: let Rel8 handle the JSON encoding/decoding of `App`s.
This makes our code a bit simpler. However, one major ramification of
this change is that if an `App` cannot be successfully decoded when
fetched from the database, this error condition will be dealt with by
throwing an out-of-band exception, which is how we handle errors
detected in the Rel8 or Hasql layers. Prior to this change, because
our code in `querySessionId` was doing the decoding, we could handle
decoding errors ourselves, and we would report these as a cryptic and
somewhat unhelpful error to the student in the UI, in-band.

In any case, this new way of handling errors as proper exceptions is
more consistent with how we handle other catastrophic errors. After
all, if a student's program can't be decoded when fetched from the
database, this probably means that something has gone horribly
wrong (a failed migration, data corruption, etc.), and should be
handled appropriately.