feat: Use our new Rel8 adapter in `primer-service`.
This change replaces the Selda database adapter with our new Rel8
adapter.

One ramification of this change is that we lose SQLite support, but
that was only ever intended as a local development option, anyway.
Recent changes (esp. Docker scripts to run a local PostgreSQL
instance) have obviated the need for it, in any case.

Error reporting on the command line leaves something to be desired,
but it's not demonstrably worse than what we had before and isn't a
priority at the moment.

Note: As of this commit, I've had to set `type-class-roots = True` in
our Weeder config. Otherwise, Weeder will complain about all of the
internal helper functions in the `Primer.Database.Rel8` namespace that
are only used via the `Rel8DbT` `MonadDb` instance. For details, see:

https://github.com/ocharles/weeder#type-class-instances