chore(ghc): Upgrade to GHC 9.2.3.
We must also bump Weeder to 2.4.0, as it tracks GHC closely.

Note that this commit includes workarounds for packages that are not
yet ready for GHC 9.2:

* protolude: ignore bounds.
* cabal-fmt & cabal-edit: build using GHC 8.10.7.

HLS is broken in by this commit. We can work around it, but it
requires some fairly major changes to the flake that are hopefully
only temporary until the next release, so I'll do that in a subsequent
commit, to make it easier to revert those changes in the future.

Somewhat unrelated, in this commit we also pin fourmolu to 0.6.0, as
the latest version (0.7.0) makes some pretty arbitrary formatting
changes *again*, and I'm tired of fixing the pre-commit checks every
time this happens.