db: Add sessions table.
Note: now that we're using Sqitch and the schema it creates is named
"primer", we need to pass the schema name to Rel8. (Previously we
passed `Nothing`, which configured Rel8 to effectively use the
PostgreSQL built-in "public" schema.)

We might later need to pass the schema name as a runtime parameter,
but we'll defer that until it's necessary.

Closes https://github.com/hackworthltd/primer/issues/242.