feat: Support importing (i.e. copying the definiton) types
The reason that srcProg and primSrcProg have been put into an
Applicative context is that later we will want to import terms, and to
test we will define some terms in these programs. This will need a
MonadFresh context. To avoid changing all use sites later, we will write
it monadically now.