fix: Prim type must be in scope to use a prim con
Since primitive constructors are hard-wired into the Expr type, they are
always in scope. However, the corresponding primitive type definitions
may not be in scope. Since we have a hard-wired map `primConName` that
gives the type any PrimCon constructs we could infer a non-existant type
for primitive constructors!

Previously, we would infer a type of `Char` for `'a'`, even when the
type `Char` did not exist in the current program, or even if it had been
arbitrarily defined by the user.

We now check that the type-name given by the hard-wired map exists in
the current program, and is a primitive type. Note that this design
relies on the fact that primitive types cannot be renamed.