fix: import: reordering case branches if necessary
Since we allow rewiring imported type dependencies to any compatible
existing type, with any compatible bijection of their constructors, we
need to handle the case where the constructors get reordered. This is
important since the typechecker enforces that the branches of a pattern
match are in the order that constructors are declared in the
corresponding type definition.

For example, if we imported and rewired the type
  data T = A | B
to
  data R = C | D
via A -> D and B -> C, then we would need to reorder the import of
  case x of
    A -> t
    B -> some_other_rhs
to
  case x of
    C -> some_other_rhs
    D -> t

(Note that much of this commit only changes indentation)