test: Use Sqitch to initialize db in pagination tests.
(Note that the pagination tests themselves didn't change here, only
their indentation.)

We remove the raw SQL in the pagination test that we previously used
to create the DB schema, and now use Sqitch, just like we would in a
real deployment.

This is a lot of code for what should be a relatively simple change,
but the implementation is complicated by a few things:

* tmp-postgres uses "keyword-value" PostgreSQL connection strings, but
Sqitch uses DB pseudo-URIs, so we have to go to some lengths to
convert between the two.

* By default, tmp-postgres sets up a UNIX socket listener for its
temporary DB, but Sqitch can't handle that, as far as I can tell, so
we need to add some extra config to tmp-postgres so that it listens on
a localhost port.

* We use tmp-postgres's database caching support, which requires quite
a bit of setup and `with`-style bracketing. We're not really taking
advantage of the caching here as we only set up and tear down the DB
once for these pagination tests, but if we ever want to add more
isolated tests to `primer-service` in the future, the caching support
I've added here should have a fairly major impact on test performance.