test(primer-rel8): Add a simple `insertSession` test.
This test simply inserts `newApp` into the database using dummy
metadata, then retrieves it and ensures that the resulting app is the
same as the one inserted.

We add an `Eq` instance to `SessionData` so that we can do the
comparison. (We also add `Show` for debugging.)

Note that, at the moment, we must override tasty's default behavior of
running tests in parallel in `primer-rel8-test`, because otherwise we
run into trouble trying to start multiple PostgreSQL instances at the
same time. There are several ways around this, but, I don't think any
of them warrant the extra effort at the moment, given the very small
number of tests we're doing.

The easiest way to fix this would be to support multiple simultaneous
PostgreSQL instances, by randomizing/hashing the port on which each
instance starts, rather than hard-coding the port number to 5234 in
`withDbSetup`.

For more complicated (but potentially more performant/less
resource-intensive) solutions, see:

https://github.com/UnkindPartition/tasty/issues/171
https://jfischoff.github.io/blog/keeping-database-tests-fast.html

However, these solutions would require more careful isolation of
tests, since they'd be running in the same database.

Tracked in:
https://github.com/hackworthltd/primer/issues/266