test(primer-rel8): Add tests for `updateSessionApp`.
Note that these tests revealed an issue with the Rel8 implementation,
namely that if you try to `updateSessionApp` using a session ID that
doesn't exist, the update will not update any rows, but this is not
detected as an error (as Rel8 will return 0 rows modified).

This should be flagged as an exception, as it means that we've done
something wrong (we should never try to update a session that doesn't
already exist in the database). However, I want to wait on
implementing this until I've had time to create a proper exception
type for Rel8/Hasql exceptions. Therefore, I've wrapped this test with
`expectFailBecause` from tasty-expected-failure for now.

See:
https://github.com/hackworthltd/primer/issues/269