feat: Add proper exceptions for `Rel8DbT` errors.
Rel8 (or rather Hasql) reports exceptions in-band via `Either`.
However, this complicates error checking for the fast path: database
errors should be truly exceptional. Therefore, we've made the choice
to throw them via `throwM`.

Prior to this change, these exceptions were generic. (I'm not even
certain what exception type they were wrapped in, as determining this
depends on which particular `throwM` instance is used in our `Rel8DbT`
transformer stack.) Now, we have a dedicated `Rel8DbException` type
with `MonadDb`-specific constructors for each type of high-level
database operation that Primer supports.

I've also changed the `primer-rel8` tests to check for the specific
exception that should occur on failure.

(FYI, the main reason we need this `Rel8DbException` type is that
Hasql doesn't make its `QueryError` an instance of `Exception`.
`Rel8DbException` basically wraps `QueryError` in an `Exception`
instance, along with the `SesssionId` of the failed operation, when
there is one.)