feat: Throw exception when a DB update session ID doesn't exist.
Row update operations in Rel8 don't fail when no rows are affected,
but in our schema, this means we've tried to update a session ID that
doesn't exist in the database. If this condition were to occur, it
would mean either that there's a bug in Primer, or that there's been a
race where the student attempted to edit a program that's been
deleted (probably due to having 2 or more browser windows open for the
same session). Previously, we did not detect this condition, but now
we do, and we throw a special exception.

We also detect when the number of affected rows on an update is > 1.
This should *never* happen, because our schema requires that session
IDs are unique, but it's easy enough to be really paranoid here and
throw another special exception when this condition is detected.

This closes https://github.com/hackworthltd/primer/issues/269.