feat: Throw exception when a DB program fails to decode.
Unlike most of our DB operations, `querySessionId` can indicate
failure in-band (via `Either`). The reason for this difference in
behavior is that the session ID provided to the query may be the
result of input from the student (directly or indirectly), and
therefore this should not be considered an error or exceptional case,
as such: it's just invalid input.

However, prior to this commit, we were also returning a `Left` value
in-band when the program couldn't be decoded properly after being
fetched from the database. This *is* an exceptional case, and
therefore this error condition should be thrown to an exception
handler. This commit implements that new behavior by throwing a new
exception value that indicates when a program decoding error has
occurred.