test(primer-rel8): Add more specific exception tests.
We now test not only that the `Rel8DbException` exception value is the
correct one, but also that the reported session ID is the expected
one.

As a result, we can rip out a number of helper functions that we're no
longer using. These might be useful in the future, but a) YAGNI and b)
they'll be easy enough to add again later.