fix: Tests.Action.Prog: defaultPrimsProg clobbers 'main'
The definition of defaultEmptyProg is a hack, hardcoding an ID of 0 for
'main' and 2 for 'other', regardless of whether they are fresh or not.
This leads to two different defs having ID 0 in defaultPrimsProg as
normally used in progActionTest, and it turns out that 'main' is
overwritten with a primitive definition. Thus one cannot do any edits to
ID 0 in tests.

Note that whilst we have worked around this symptom, the implementation
of defaultEmptyProg is still wrong: these IDs will clash with some node
IDs in the current usage.