feat!: Progs can contain imported modules
We enable Primer programs to import modules. This is achieved by adding
a `progImports :: [Module]` field to `Prog`. The imported modules are
each immutable (via the api). This immutability is ensured by having all
the actions only operate on the one "current" module (that comprises the
old `progDefs` and `progTypes`). (For now, the set itself is also
immutable via the api.)

For now we don't have any namespacing, we just lookup references in all
modules and assume unique names. We leave adding hierarchical names for
future work. We also leave having multiple editable modules for future
work.

BREAKING CHANGE: this change requires a database migration, as it
changes the representation of `Prog`. However, since this is just
serialised to json and stored as a blob in the DB, it requires no schema
changes. Since we have no programs we need to preserve, we decided not
to bother with a migration. This means that DBs created before this
commit will not load with a primer containing this commit.