test(primer-rel8): Get the PostgreSQL port dynamically.
This will permit us to run multiple PostgreSQL instances
simultaneously.

Note that there's a theoretical race condition in this implementation,
but it's extremely unlikely to be triggered.

Also note: the `tmp-postgres` docs claim that if you use its default
configuration, it will listen on localhost on a random port, configure
a username and password, etc., but I could not get this to work.