test(primer-rel8): We should now be able to run DB tests in parallel.
There's even more we could do to increase parallelism, because most of
these tests don't trample on each other. (The one exception I can
think of is the `test_listSessions` test, which depends on the state
of the database being just so and therefore needs to run on a
clean (and isolated) database.)

However, in order to get more parallelism, we'd need to set up a
database connection pool, and we'd need to re-organize the tests so
that we can wrap all the parallelizable tests with a single
`withDbSetup` bracket. For the moment, the tests are already pretty
fast, so let's defer that extra work and complexity until we need it.

Closes https://github.com/hackworthltd/primer/issues/266