test: Add a database initialization CI test.
This test runs `sqitch deploy` directly against a fresh test database,
using the latest schema. No Primer backend code is involved at all.

Note that this test is intended for CI only, mainly because it sets
up/tears down the test database using a Buildkite Docker plugin. It
uses a custom Buildkite pipeline and is not part of our standard Nix
test suite and CI pipeline. With a bit of work, we could make it part
of the standard developer test suite, but it's probably not worth it:
consider it a deployment test, and not something that Primer
developers would typically need to worry about.