test: Add a database initialization test.
This test runs `sqitch deploy` directly against a fresh test database,
using the current Primer database schema. No Primer backend code is
involved at all: this is purely Sqitch and PostgreSQL. This test could
therefore be considered to be a unit test for our schema.

This test uses the `nixpkgs` NixOS test harness. It boots a VM running
the NixOS PostgreSQL service, then runs Sqitch to deploy our schema
and verify that the deployment worked. This is probably quite a bit
different than what we'll eventually use to deploy our schema to a
production database, but as we don't yet know what that production
PostgreSQL service will be, this will suffice for now.

One alternative to the NixOS VM approach would be to run a Docker
container with the official PostgreSQL image. That would more closely
resemble our eventual production service (assuming we use a container
solution, which looks very likely), but the Docker approach has a few
downsides:

1. It would require quite a bit of work to make the Docker approach
work both in CI and from a developer machine interactively. (On the
other hand, the eventual production-like test will also probably only
run in CI, so perhaps this isn't much of a concern after all.)

2. Subsequent commits to this one will add support for `pgtap`, which
is a PostgreSQL-based unit test system. `pgtap` is a PostgreSQL
extension, and getting extensions to work with the official PostgreSQL
Docker container is non-trivial, whereas in NixOS, it's just a
one-line config change.

In any case, the NixOS solution is more expedient and will do for the
time being.