test: Add a simple unit test for the `sessions` table.
Here we use the `pgtap` PostgreSQL extension and its accompanying
`pg_prove` utility to run a simple unit test against the `sessions`
table. This test isn't particularly interesting on its own, but it's a
good PoC for the idea of PostgreSQL unit testing with this framework.

As with the previous commit, we use the NixOS test framework here to
run the test. This is much easier than trying to get `pgtap` installed
into the official PostgreSQL Docker image, but it's probably a very
different environment than what we'll use in production, so this is
really just a stopgap/convenience until we have a proper deployment
system that we can test.