feat(primer-rel8)!: Encode `App`s as JSONB in PostgreSQL.
This change will enable many potential optimizations for fetching
sub-trees from an `App`s program, because PostgreSQL can perform
path-based JSON operations on JSONB columns. In addition, it should be
a more efficient format for storing our programs. See
https://www.postgresql.org/docs/13/datatype-json.html for details.

Note that this change may have implications on what names we can
encode in the database. The previous `App` encoding was a raw binary
encoding, so string encodings were irrelevant to the database. Now
that we're using JSONB, the database must be able to interpret strings
in the encoding, and therefore, we can only encode strings that are
compatible with the database's own encoding. Our database schema
expects that the database is configured to use UTF8, so we *should* be
OK, but I confess that there may be some Unicode corner cases where we
accept some names that will not be compatible with UTF8 JSONB.
Presumably, these insertions/updates will fail and we'll be able to
catch any such corner cases as they arise, but it probably warrants
some testing in the meantime. See:
https://github.com/hackworthltd/primer/issues/287.

This change removes the `LoadSessionProgramDecodingError` value in the
`Rel8DbException` type. We used to be able to throw that error
specially, because it was our code that was doing the Aeson decoding
of the `App` we retrieved from the database. Now, however, the
decoding is done at the Rel8 level (via the `JSONBEncoded` instance),
and therefore we can no longer generate a special exception when the
program doesn't decode properly. Presumably such errors will now
manifest via one of our generic Rel8/Hasql exceptions. As a result,
I've also removed the test case that used to exist for this condition,
because now it's much more difficult to insert a row with an invalid
program (from Haskell, anyway). In any case, these decoding errors
should effectively never happen, short of the database being
out-of-sync with the app, and this is a "category 3" error (in our new
parlance), meaning that the student can't do anything about it,
anyway. Therefore, the fact that program decoding errors are now
reported generically as a Rel8/Hasql exception should not cause any
problems.

BREAKING CHANGE: this change requires a database migration. If you
want to preserve your existing PostgreSQL development database, you
must run the `deploy-local-db` script against it. Assuming your
PostgreSQL Docker container is running, run this:

nix run .#deploy-local-db

Closes https://github.com/hackworthltd/primer/issues/103
Closes https://linear.app/hackworthltd/issue/PRIM-4/proper-migration-system