feat: `importModules` utility; refactor TC modules
This is basically an action, except we don't expose it in the API. We
will delay doing so until we have a nice way for a frontend to refer to
a module. I.e. we would like to say
"import the User1.ProjectFoo.ModuleBar" module, rather than having to
explicitly give the code contained within the module.

This utility does not check that imported modules have disjointly-named
contents. We intend to shortly require a hierarchical naming scheme
which will address this concern. At that point we will also address the
issue of disjoint IDs of contents, and our freshness guarantees.

The reason we delay worrying about disjointness/uniqueness is that both
it and being able to expose this nicely in the API require the notion of
a module having some short name, rather than having to explicitly pass
around all the code contained in the module.