DNM: Testing CI + golden tests: accepts non-existent golden files?
We add a golden test fixture but do not check in its output file.
Running locally, I see
  $ cabal run primer-test -- -p 'golden'
  Tests
    Serialization
      encode
        golden_file_not_checked_in: OK
          Golden file did not exist; created
      decode
        golden_file_not_checked_in: FAIL
          Exception: test/outputs/serialization/golden_file_not_checked_in.json: openBinaryFile: does not exist (No such file or directory)
          Use -p '/golden/&&/decode.golden_file_not_checked_in/' to rerun this test only.
i.e. the testsuite only fails because of the decode test.
The encode test (using tasty-golden) passes, even though the golden file
did not exist.

I am worried that this could lead to situations where we forget to check
in a golden file, and this is not caught by CI. However, in our current
testsuite, I think we are saved by the fact the only golden tests we do
are via Fixtures, which include this decode component which will fail.