refactor: Remove IDs from definitions
We now identify them by `Name` instead.

The original motivation for using IDs was to make renaming easy, as mentioned in hackworthltd/vonnegut@7d88864. But we eventually decided to take a traditional approach to name shadowing, so we have to ensure that names don't clash anyway, and thus lose much of the benefit of using IDs, but continue to carry around both a Name and an ID for little reason. Furthermore, we never applied the same logic to other constructs such as TypeDefs, to which we don't assign IDs. So in this commit we simplify everything by going back to using IDs only for their original purpose of identifying things which don't have names, i.e. subexpressions.

For now, the `RenameDef` action is broken, since we don't update references to the renamed definition. This will be fixed in the subsequent commit, as we'll be able to use the `renameVar` function which currently only works for local variables.