refactor: disambiguate type constructor names
We use a new type TyConName to refer to type constructors, rather than
Name. This is the start of a broader refactor to avoid using Name for
everything, to clarify what sort of name each is.

To avoid changing the API, our ToJSON, FromJSON and ToSchema instances
are identical to the ones for Name.