refactor: disambiguate local variable names
We clarify which names are (expected to be) locally bound. We treat all
locally-bound things in the same way (various sorts of `let`,
lambda-bound variables etc), as we rarely care about their differences.
(In contrast to always caring whether a name refers to a type
constructor or a local variable.)