test: types of our primitive defs are well formed
This requires some support in the typechecker. We add it to the checks
we do when importing a module. Since these primitives are hardcoded,
this should never fail in deployment, but it should catch (in CI) any
mistakes made when adding further primitives in the future, assuming we
add the primitive to the default 'newProg'.

Note that the main impact of this commit is to enhance the existing
'unit_good_defaults', which calls 'checkEverything'.