fix: make typeDef well formed in Tests/Serialization
We were accidentally using TCon to refer to a type parameter.
Whilst this test does not care whether things are well formed, this is
useful preparation for making global names qualified with a module name.
When that happens, the contents of a TCon will be a qualified name, but
the contents of a TVar will be a non-qualified (local) name,
necessitating a change here.